Code/inference/requirements.txt
numpy==1.25.2
Pillow==9.4.0
gradio==4.28.3
matplotlib==3.7.1
pandas==2.0.3
nltk==3.8.1
keras==2.15.0
h5py==3.9.0
